rename the directory to use pascal case
