+++ Model/csvService/classMap/UserClassMap.cs
@@ -1,7 +1,7 @@
 ﻿using AirportTicketBooking.Model.Classes;
 using CsvHelper.Configuration;
 
-namespace AirportTicketBooking.Model.csv_service.Class_map;
+namespace AirportTicketBooking.Model.csvService.classMap;
 
 public class UserClassMap: ClassMap<User>
 {
