+++ Model/csvService/csvReaders/CsvFlightReader.cs
@@ -1,9 +1,9 @@
 ﻿using AirportTicketBooking.Model.Classes;
-using AirportTicketBooking.Model.csv_service.Class_map;
+using AirportTicketBooking.Model.csvService.classMap;
 using CsvHelper;
 using System.Globalization;
 
-namespace AirportTicketBooking.Model.csv_service.Csv_Readers;
+namespace AirportTicketBooking.Model.csvService.csvReaders;
 
 public class CsvFlightReader : ICsvReader<Flight>
 {
